Seperate role tests from members
